+++ src/HoneybeeDotNet/Model/ScheduleFixedIntervalAbridged.cs
@@ -44,9 +44,9 @@ namespace HoneybeeDotNet.Model
         /// <param name="scheduleTypeLimit">Name of a ScheduleTypeLimit that will be used to validate schedule values against upper/lower limits and assign units to the schedule values. If None, no validation will occur..</param>
         /// <param name="timestep">An integer for the number of steps per hour that the input values correspond to.  For example, if each value represents 30 minutes, the timestep is 2. For 15 minutes, it is 4. (default to 1).</param>
         /// <param name="startDate">A list of two integers for [month, day], representing the start date when the schedule values begin to take effect.A third integer may be added to denote whether the date should be re-serialized for a leap year (it should be a 1 in this case)..</param>
-        /// <param name="placeholderValue"> A value that will be used for all times not covered by the input values. Typically, your simulation should not need to use this value if the input values completely cover the simulation period. (default to 0M).</param>
+        /// <param name="placeholderValue"> A value that will be used for all times not covered by the input values. Typically, your simulation should not need to use this value if the input values completely cover the simulation period. (default to 0).</param>
         /// <param name="interpolate">Boolean to note whether values in between intervals should be linearly interpolated or whether successive values should take effect immediately upon the beginning time corrsponding to them. (default to false).</param>
-        public ScheduleFixedIntervalAbridged(string name, List<decimal> values, string type = "ScheduleFixedIntervalAbridged", string scheduleTypeLimit = default, int timestep = 1, List<int> startDate = default, decimal placeholderValue = 0M, bool interpolate = false)
+        public ScheduleFixedIntervalAbridged(string name, List<double> values, string type = "ScheduleFixedIntervalAbridged", string scheduleTypeLimit = default, int timestep = 1, List<int> startDate = default, double placeholderValue = 0, bool interpolate = false)
         {
             // to ensure "name" is required (not null)
             if (name == null)
@@ -91,7 +91,7 @@ namespace HoneybeeDotNet.Model
             // use default value if no "placeholderValue" provided
             if (placeholderValue == null)
             {
-                this.PlaceholderValue = 0M;
+                this.PlaceholderValue = 0;
             }
             else
             {
@@ -122,7 +122,7 @@ namespace HoneybeeDotNet.Model
         /// <value>A list of timeseries values occuring at each timestep over the course of the simulation.</value>
         [DataMember(Name="values", EmitDefaultValue=false)]
         [JsonProperty("values")]
-        public List<decimal> Values { get; set; }
+        public List<double> Values { get; set; }
 
         /// <summary>
         /// Gets or Sets Type
@@ -161,7 +161,7 @@ namespace HoneybeeDotNet.Model
         /// <value> A value that will be used for all times not covered by the input values. Typically, your simulation should not need to use this value if the input values completely cover the simulation period.</value>
         [DataMember(Name="placeholder_value", EmitDefaultValue=false)]
         [JsonProperty("placeholder_value")]
-        public decimal PlaceholderValue { get; set; }
+        public double PlaceholderValue { get; set; }
 
         /// <summary>
         /// Boolean to note whether values in between intervals should be linearly interpolated or whether successive values should take effect immediately upon the beginning time corrsponding to them.
